Prettify column defaults (#44)
Fixes #45. 

When I extracted out column annotation building [1], I changed the way
the default value was read. `Model#column_defaults` has the column
default values that are casted appropriately, where as `Column#default`
is uncasted and straight from the database, most of the time as a
String. It's important to note that the various adapters (sqlite, mysql,
postgres) implement the columns differently [2], especially as it
relates to Column#default and Model#column_default values.

This PR restores parity with the original gem on how defaults are
handled. It also adds tests for the various adapters, because the
implementations vary under the hood.

As an example, the mysql adapter sets the value for an integer field
without a default to be `0` in `Model#column_defaults`, where as sqlite
and postgres set it to be `nil`.

[1]
https://github.com/drwl/annotaterb/pull/28/files#diff-92c15977800c36cb911c1d7971ede0da49dae93b0c9e555777a5ca67ee81d93dR12-R14

[2] https://gist.github.com/drwl/ef274a763b773d132abfdb95d32347d2