Support non-model files in `CustomParser` (#88)
This PR adds support for other Ruby files types that should be able to
get model annotations added to them (i.e. factories, fabricators, etc).

In #72, AnnotateRb was changed to use `FileParser::CustomParser` to
parse Ruby files instead of relying on regexes. That PR only added
support for model files that used either `module Namespace...` or `class
ModelName...` as the first line of Ruby. The parser did not have support
for other file structures like FactoryBot factories or Fabrication
fabricators.

There may be other Ruby files that have code that is not currently
handled by `CustomParser`, so those will have to be added in the future.