Change AnnotationDecider to return false early (#112)
In the old gem [1], the conditional for deciding to annotate a file was
chained, meaning that execution of further checks stopped upon reaching
a false condition.

The conditional was hard to read, so I decomposed the conditional into
readable chunks. This led to a regression, that deciding to annotate a
`Module` would raise because `#superclass` is a method on classes, but
not for instances of modules. By returning false early, we fix the
regression.

To illustrate, before this change one would see this in the output:

```
Unable to process app/models/user_settings/dsl.rb: undefined method `superclass' for module UserSettings::DSL
```

when attempting to annotate a file like:
```ruby
# app/models/user_settings/dsl.rb

module UserSettings::DSL
  module ClassMethods
    # ... methods ...
  end
end
```

In the old gem, this error message wouldn't be outputted, so this change
makes AnnotateRb match the behavior.

[1]
https://github.com/ctran/annotate_models/blob/e60a66644e8a8ea5dccd6a398d31f22878233998/lib/annotate/annotate_models.rb#L734-L738