Make `exclude_tests` option able to override other exclude options (#107)
This PR adds support for `exclude_tests` option to be a boolean or an
array. When it's a boolean, the behavior should stay the same. When it's
an Array, any of the following options can be included:

Prior to this change, `exclude_scaffolds` would used for controller
tests, request specs, and routing specs. With this change, it will
depend on `exclude_scaffolds` AND `exclude_tests` assuming
`exclude_tests` is an array.

```ruby
:model, :controller, :serializer, :request, :routing
```

Because it's an exclusion list, options being present will mean to
exclude them. Where as the absence of an option means to include it.
Options specified or not specified here will take precedence over their
original option; more explained below.

_I think this is very confusing so changing the `exclude_` options to
become `include_` should be something to consider for the future_

To illustrate the behavior:
When `exclude_tests: [:model]` => this means to include controller
tests, serializer tests, request specs, and routing specs. This will
take precedence over `exclude_scaffolds: true`.

When `exclude_tests: [serializer] and exclude_serializers: false`, this
means that serializers will be included but their tests will be
excluded.

**Note: when setting exclude_tests to an Array, to use the singular form
and not the plural form.**

For example, to exclude controller tests:
```yaml
:exclude_tests: [controller]
```