Refactor model annotation components (#134)
This PR is mostly a pure refactor, in that it does not change the
behavior of any public methods. The only exceptions are on
`ModelAnnotator::AnnotationBuilder`, with `#header`,
`#schema_header_text`, and `#schema_footer_text` being inlined or their
logic being extracted.

It introduces a new pattern for generating parts of a model annotation,
with the preference for separating presentation from data fetching. As a
person working on AnnotateRb, it should also be easier to understand
when there are format specific differences (i.e. markdown vs default).