Switches to the block form of String#sub
There are a number of different mechanisms with String#sub that pertain
to the matching of backreferences and capture groups when using either
the String or Hash type args for the replacement arg.

When an index contains unescaped characters, those are being incorrectly
interpreted as capture groups.

>If replacement is a String that looks like a pattern's capture group
but is actually not a pattern capture group e.g. "\'", then it will have
to be preceded by two backslashes like so "\\'".

Using the block version of #sub avoids this.