Support parsing of dynamic fixture erb yml files (#158)
This PR adds support for dynamic fixtures, which are yml files with Ruby
code in them. Rails supports this so we also want to support this.

Before this change, Ruby's yml parser, Psych, would raise
`Psych::SyntaxError` when attempting to parse non-valid yml. This change
attempts to parse yml normally, but if it fails will run the input
through the Erb evaluator and then attempt to parse the output using
Psych.

Addresses #149