Fix translation foreign key exclusion bug (#181)
I ran into a bug with this logic when using annotaterb + globalize:

For compound names for models such as "ABCustomer", it resulted in
"abcustomer_id" being added to the exclusion list, where the actual
foreign key name was "ab_customer_id".

Therefore it was not excluded. And therefore all the annotations for
ABCustomer ended up including "ab_customer_id" even though that wasn't
an actual column.

This PR includes changes to the unit tests to not break them. However it
didn't seem like this bug something I could actually reproduce with the
unit tests. I couldn't get the integration test suite working and
couldn't spend more time on it right now.

---------

Co-authored-by: Andrew W. Lee <git@drewlee.com>