Add expected file to automatically require (#185)
Due to Bundler's naming conventions, `annotaterb` gem does not get
automatically required. Per https://github.com/drwl/annotaterb/pull/177,
it's expected to have a corresponding `lib/annotaterb.rb` file, but it
does not.

This fixes it so that it automatically gets required.

Credit to @kbruccoleri for reporting and @yskkin for suggesting the fix.