show included columns in indexes (#211)
This PR succinctly adds support for showing columns included in an index
using `INCLUDE`. This is a thing in, at least, PostgreSQL and MSSQL.

Also, in order to run the integration tests, I had to add some gems
explicitly to the dummyapp Gemfile as they're not part of the stdlib as
of Ruby 3.4.0 - let me know if this is not desirable.

---------

Co-authored-by: Andrew W. Lee <git@drewlee.com>