Place column comments at the end of the line (#199)
Fixes #164 

Adds the configuration `position_of_column_comment`, now defaults to
`with_name` (already existing behavior), and can also be set to
`rightmost_column`.

There was a bit of moving code around involved, as some information
about other attribute was needed (e.g.: the `attributes` of other
columns, to determine how wide this column must be).

The only formats where it got implemented though are:

- default (text)
- Markdown

The other formats I do not know well, and would ask anyone with better
knowledge of them to work on it.

---------

Co-authored-by: Andrew W. Lee <git@drewlee.com>