fix: use model name for file retrieval when not connected to the primary DB. (#207)
When the same table name exists in different databases, attempting to
retrieve fixture and factory files references them by the table name. As
a result, in addition to the fixtures and factories of models connected
to the secondary database, related files from the primary database are
also retrieved.

Therefore, this commit modifies the behavior so that for models
connected to the secondary database, related files are retrieved based
on the model name instead of the table name.

fixes #206

Co-authored-by: Andrew W. Lee <git@drewlee.com>