Add AnnotateRb::Runner.running? method (#242)
This is so that we can facilitate dynamic logic during the annotation
process.

This is something that the historical `annotate_models` gem provided,
via `Annotate::Migration.class_variable_get(:@@working))`


https://github.com/ctran/annotate_models/blob/5d01c4171990c4fe7b9b0977b05ce14a98209e53/lib/tasks/annotate_models_migrate.rake#L36-L41

Instead of having to reach into the Migration class, this instead stores
some state on the `AnnotateRb::Runner` class so that we can ask it
directly if it's in the middle of running.