Speed up AnnotationDecider (#243)
While looking for a fix for #172 I noticed that `AnnotationDecider` does
a fair amount of checks before checking that _all_ are true. We can
speed up the decision making a bit and avoid the current issue with
`AssociatedObject` completely if we return false early whenever a check
fails instead of collecting all the results and then returning false.
It's not the prettiest code but it works and is faster (slightly) on a
benchmark:

```
Comparison:
NewAnnotationDecider#annotate?:          30189.7 i/s
OriginalAnnotationDecider#annotate?:    18295.3 i/s - 1.65x  slower
```

I'm aware that the issue posted in #172 Is actually an issue with the
`AssociatedObject` gem so no worries if you don't want to merge this.

---------

Co-authored-by: Kasper Timm Hansen <hey@kaspth.com>