Fix STI models not being annotated (#256)
This PR fixes STI classes not being annotated. This happened because in
https://github.com/drwl/annotaterb/pull/243, the
`descends_from_active_record?` check happens before checking
`@options[:exclude_sti_subclasses]`.

Before the change in #243, it was using an old check to determine if a
model was using STI [1]. Rails added the `descends_from_active_record?`
that will:
- return `true` if it is using STI [2], which requires a `type:string`
field and the model be subclassed
- return `false` otherwise

so we're switching to use that instead.

[1]
https://github.com/drwl/annotaterb/pull/243/files#diff-fac5363b884746fd38e2047fb25e19daff331d7dca2d6fe4b8b4fcc259791f3bL29
[2]
https://guides.rubyonrails.org/association_basics.html#single-table-inheritance-sti

Fixes https://github.com/drwl/annotaterb/issues/252