Escape column comment newlines when using "rightmost" value for position_of_column_comment option (#263)
This is a follow-up to the changes added in
https://github.com/drwl/annotaterb/pull/199.

We noticed that when using the `rightmost` option for column comment
position, any multi-line column comments were not being escaped in the
same manner as the `with_name` option. This causes problems with
generated model annotation as invalid uncommented newlines then appear.
For example:

**Before**

```
#  notes                                        :text(55)         not null     Notes.
May include things like notes.
```

**After**

```
#  notes                                        :text(55)         not null     Notes\nMay include things like notes.
```

I have also updated the documentation for this option to include the
available values – I had to dig through the codebase to find the
possible values for the `--position-of-column-comments` option.

Thanks for this great gem, my company uses it in our projects 👏