Support .config directory convention (#281)
In an effort to reduce the number of configuration dotfiles from the
project root, I came across the `.config/` convention. RuboCop already
supports this and it would be nice for this library to as well:
https://docs.rubocop.org/rubocop/configuration.html#config-file-locations

This PR allows you to put your config files in the following location
(ordered by precedence):
- `.annotaterb.yml` (default)
- `config/annotaterb.yml` (Rails friendly)
- `.config/.annotaterb.yml` (https://dot-config.github.io)
- `.config/annotaterb/config.yml` (https://dot-config.github.io)

More reading about this movement: https://dot-config.github.io

I've also taken the liberty of cleaning up somethings in this PR as
well.